requirements.txt
matplotlib==3.8.2
numpy==1.26.2
Pillow==10.1.0
torch==2.1.2
torch_summary==1.4.5
torchvision==0.16.2
tqdm==4.66.1
